refactor(api-file-manager-s3): DI-native S3 GraphQL schema factory
Convert the S3 GraphQL schema from a legacy createGraphQLSchemaPlugin
(service-locating context.container.resolve inside resolvers, bridged via
addLegacyResolvers) to a CoreGraphQLSchemaFactory that declares each resolver's
dependencies and uses builder.addResolver. IdentityContext/GetSettingsUseCase/
TenantContext are now injected; env-based S3 client + multipart use-cases stay
inline. Behavior-preserving; first slice of the GraphQLSchemaPlugin -> DI
schema-contributor migration.
